+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix: folder markdown about_page to about-page
 - fix: demo image path in README
+- refactor: replace app_version meta tag with `__APP_VERSION__` constant
 
 ## 0.12.9 (2024-09-23)